+++ src/package/implementations/PlayerImpl.ts
@@ -1,4 +1,4 @@
-import { IPlayer } from "../interfaces/Player";
+import { IPlayer } from "../interfaces/IPlayer";
 
 class PlayerImpl implements IPlayer {
