+++ src/package/interfaces/IPlayer.ts
@@ -1,0 +1,1 @@
+export interface IPlayer {}
