+++ src/package/interfaces/IRoomEvent.ts
@@ -1,0 +1,3 @@
+export interface IRoomEvent {
+    getType(): string;
+}
