--- src/package/interfaces/Player.ts
@@ -1,3 +1,0 @@
-export interface IPlayer {
-
-}
